requirements.txt
requests
beautifulsoup4
opencc-python-reimplemented
